Create Service for CurvePoint and Correction Repos

+++ P7CreateRestApi/Domain/CurvePoint.cs
@@ -3,7 +3,7 @@ namespace Dot.Net.WebApi.Domain
     public class CurvePoint
     {
         public int Id { get; set; }
-        public byte? CurveId { get; set; }
+        public int? CurveId { get; set; }
         public DateTime? AsOfDate { get; set; }
         public double? Term { get; set; }
         public double? CurvePointValue { get; set; }
