+++ P7CreateRestApi/Domain/DTO/CurvePointDTO.cs
@@ -3,7 +3,7 @@
     public class CurvePointDTO
     {
         public int Id { get; set; }
-        public byte? CurveDTOId { get; set; }
+        public int? CurveDTOId { get; set; }
         public DateTime? AsOfDate { get; set; }
         public double? Term { get; set; }
         public double? CurvePointDTOValue { get; set; }
